+++ src/core/lockfile.rs
@@ -18,7 +18,8 @@
 use crate::core::ip::Ip;
 use crate::core::manifest::FromFile;
 use crate::core::manifest::IpName;
-use crate::core::source::Source;
+use crate::core::source;
+use crate::core::source::Repository;
 use crate::core::uuid::Uuid;
 use crate::core::{catalog::CacheSlot, ip::IpSpec};
 use crate::core::{
@@ -280,8 +281,8 @@ pub mod v1 {
         uuid: Uuid,
         // @note: `sum` is optional because the root package will have its sum omitted
         checksum: Option<Sha256Hash>,
-        #[serde(flatten)]
-        source: Option<Source>,
+        #[serde(deserialize_with = "source::read_string", default)]
+        repository: Option<Repository>,
         // @note: `path` is optional and only used if the dependency list uses a local ip
         path: Option<PathBuf>,
         dependencies: Vec<PartialIpSpec>,
@@ -307,7 +308,7 @@ pub mod v1 {
                 } else {
                     None
                 },
-                source: target.get_man().get_ip().get_source().cloned(),
+                repository: target.get_man().get_ip().get_source().clone(),
                 dependencies: match target.get_man().get_deps_list(is_local, true).len() {
                     0 => Vec::new(),
                     _ => {
@@ -387,7 +388,7 @@ pub mod v1 {
                 } else {
                     None
                 },
-                source: ip.get_man().get_ip().get_source().cloned(),
+                repository: ip.get_man().get_ip().get_source().clone(),
                 dependencies: match ip.get_man().get_deps_list(is_working, true).len() {
                     0 => Vec::new(),
                     _ => {
@@ -529,8 +530,8 @@ pub mod v1 {
             &self.uuid
         }
 
-        pub fn get_source(&self) -> Option<&Source> {
-            self.source.as_ref()
+        pub fn get_source(&self) -> Option<&Repository> {
+            self.repository.as_ref()
         }
 
         pub fn get_name(&self) -> &IpName {